Updated documentation, README.md, .gitignore and fixing an error with JSON
