+++ .gitignore
@@ -26,3 +26,4 @@ yarn-debug.log*
 yarn-error.log*
 
 *aws_access*
+*statistics*
